fix(claim manager): make poolId uitn128 + add & remove claims in pool factory

+++ contracts/interfaces/IClaimManager.sol
@@ -2,17 +2,16 @@
 pragma solidity ^0.8;
 
 interface IClaimManager {
-  function getProtocolAgreement(uint256 poolId)
-    external
-    view
-    returns (string memory);
+  function getProtocolAgreement(
+    uint128 poolId
+  ) external view returns (string memory);
 
   function claimInitiator(uint256 disputeId_) external view returns (address);
 
   function claimChallenger(uint256 claimId_) external view returns (address);
 
   function addAgreementForProtocol(
-    uint256 poolId_,
+    uint128 poolId_,
     string calldata ipfsAgreementCid_
   ) external;
 }
